Move toolbar components to `src/h5web/toolbar` folder
