+++ src/h5web/dataset-visualizer/mapper/AxisMapper.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement } from 'react';
 import { isNumber } from 'lodash-es';
-import ToggleGroup from '../../visualizations/shared/ToggleGroup';
+import ToggleGroup from '../../toolbar/controls/ToggleGroup';
 import styles from './DimensionMapper.module.css';
 import type { Axis, DimensionMapping } from '../models';
