+++ src/h5web/toolbar/OverflowMenu.module.css
@@ -4,15 +4,15 @@
 }
 
 .btn {
-  composes: btn from '../shared/Toolbar.module.css';
+  composes: btn from './Toolbar.module.css';
 }
 
 .btnLike {
-  composes: btnLike from '../shared/Toolbar.module.css';
+  composes: btnLike from './Toolbar.module.css';
 }
 
 .icon {
-  composes: icon from '../shared/Toolbar.module.css';
+  composes: icon from './Toolbar.module.css';
 }
 
 .menu {
