+++ src/h5web/toolbar/controls/ColorMapSelector.module.css
@@ -5,11 +5,11 @@
 }
 
 .btn {
-  composes: btn from '../shared/Toolbar.module.css';
+  composes: btn from '../Toolbar.module.css';
 }
 
 .btnLike {
-  composes: btnLike from '../shared/Toolbar.module.css';
+  composes: btnLike from '../Toolbar.module.css';
   padding: 0 0.375rem 0 0.625rem;
 }
