+++ src/h5web/toolbar/controls/DomainSlider.module.css
@@ -31,7 +31,7 @@
 }
 
 .thumbBtnLike {
-  composes: btnLike from '../shared/Toolbar.module.css';
+  composes: btnLike from '../Toolbar.module.css';
   min-width: 2rem;
   padding: 0.375rem;
   background-color: white;
@@ -43,12 +43,12 @@
 }
 
 .resetBtn {
-  composes: btn from '../shared/Toolbar.module.css';
+  composes: btn from '../Toolbar.module.css';
   padding: 0 0.125rem;
 }
 
 .resetBtnLike {
-  composes: btnLike from '../shared/Toolbar.module.css';
+  composes: btnLike from '../Toolbar.module.css';
 }
 
 .resetIcon {
