+++ src/h5web/toolbar/controls/SnapshotButton.tsx
