+++ src/h5web/toolbar/controls/ToggleBtn.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import type { IconType } from 'react-icons';
-import styles from './Toolbar.module.css';
+import styles from '../Toolbar.module.css';
 
 interface Props {
   label: string;
