+++ src/h5web/toolbar/controls/ToggleGroup.module.css
@@ -4,12 +4,12 @@
 }
 
 .btn {
-  composes: btn from '../shared/Toolbar.module.css';
+  composes: btn from '../Toolbar.module.css';
   padding: 1px;
 }
 
 .btnLike {
-  composes: btnLike from '../shared/Toolbar.module.css';
+  composes: btnLike from '../Toolbar.module.css';
 }
 
 .btn:not(:first-child) > .btnLike {
@@ -31,9 +31,9 @@
 }
 
 .icon {
-  composes: icon from '../shared/Toolbar.module.css';
+  composes: icon from '../Toolbar.module.css';
 }
 
 .label {
-  composes: label from '../shared/Toolbar.module.css';
+  composes: label from '../Toolbar.module.css';
 }
