--- src/h5web/visualizations/heatmap/SnapshotButton.module.css
@@ -1,21 +1,0 @@
-.link {
-  composes: btn from '../shared/Toolbar.module.css';
-  color: inherit;
-  text-decoration: none;
-}
-
-.btnLike {
-  composes: btnLike from '../shared/Toolbar.module.css';
-}
-
-.icon {
-  composes: icon from '../shared/Toolbar.module.css';
-}
-
-.label {
-  composes: label from '../shared/Toolbar.module.css';
-}
-
-.link:hover > .btnLike {
-  background-color: var(--secondary-light);
-}
